moved tests to separate files based on class
