+++ src/memefeed/reddit/test_search_compose.py
@@ -1,0 +1,5 @@
+"""
+Test compose search page.
+"""
+
+# TODO:
